requirements.txt
streamlit==1.28.1
google-generativeai>=0.7.0
pypdf==4.0.1
Pillow==10.0.1
numpy>=1.24.3
python-dotenv==1.0.0
